Making Config class immutable
