+++ build.sbt
@@ -90,7 +90,7 @@ lazy val wvletConfig =
     description := "wvlet configuration module",
     libraryDependencies ++= Seq(
       "org.yaml" % "snakeyaml" % "1.14",
-      "org.wvlet" %% "airframe" % "0.3" % "test"
+      "org.wvlet" %% "airframe" % "0.6" % "test"
     )
   ).dependsOn(wvletObj, wvletTest % "test->compile")
